requirements.txt
pandas
uuid
python-dotenv
mysql-connector-python
ydata-profiling
pymysql
seaborn
matplotlib
